requirements.txt
googletrans_py==4.0.0
python-telegram-bot==20.2
